UPSTREAM: media: rc keymaps: add SPDX identifiers to the code I wrote
As we're now using SPDX identifiers, on the several
RC keymap files I wrote, add the proper SPDX, identifying
the license I meant.

As we're now using the short license, it doesn't make sense to
keep the original license text.

Also, fix MODULE_LICENSE to identify GPL v2, as this is the
minimal license requirement for those modles.

Change-Id: I858bde57da1269b6d6d92a868e051bb0ee2e06d8
Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>
Signed-off-by: Edwiin Kusuma Jaya <kutemeikito0905@gmail.com>