sched/fair: Don't allow SchedTune boosted tasks to be migrated to small cores
We want boosted tasks to run on big cores. But CAF's load balancer
changes do not account for SchedTune boosting, so this allows for
boosted tasks to be migrated to a suboptimal core. Let's mitigate
this by setting the LBF_IGNORE_BIG_TASKS for tasks migrating from a
larger capacity core to a smaller one and to check if the task is
SchedTune boosted. If both are true, do not migrate this task.

Signed-off-by: Zachariah Kennedy <zkennedy87@gmail.com>
Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Change-Id: Iacd6a06e5df06eeef587e70a7b0b8cbb876672b9
Signed-off-by: Edwiin Kusuma Jaya <kutemeikito0905@gmail.com>