qcacld-3.0: Rectify improper typecast in hdd_report_max_rate
The variable or_leng in the function hdd_report_max_rate is declared as
a 8 bit unsigned int. But while passing it to the function
sme_cfg_get_str, the value is being typecasted into 32 bit. This lower
to higher explicit typecast can result in the adjacent memory location
to be corrupted.

To ensure that there is no such scenario, change the type of the
variable or_leng to unsigned 32bit int.

Change-Id: I26dc7d4a5b6b98024c65649bc381f39c4dcde1d9
CRs-Fixed: 2487027