Revert "msm: vidc: Inline HAL_FLIP value with omx counterpart"
This break screenshot taking, by generating an image with
the wrong metadata.

The commit mentions that this patch inlines the available
image flip values with the OMX component, but the only
place I could find similar values defined are display HAL,
and they followed the old values.

enum struct DRMRotation {
  FLIP_H = 0x1,
  FLIP_V = 0x2,
  ROT_180 = FLIP_H | FLIP_V,
  ROT_90 = 0x4,
};

Bring back the old values to fix this.

This reverts commit 6601e4f5a2a2ce0040291b8ab2c9893e634a2e9f.

Signed-off-by: Fiqri Ardyansyah <fiqri15072019@gmail.com>
Signed-off-by: Edwiin Kusuma Jaya <kutemeikito0905@gmail.com>